fix authorization for all controller + add error message

+++ src/Controller/UserController.php
@@ -24,11 +24,12 @@ final class UserController extends AbstractController
         SerializerInterface $serializer
         ): JsonResponse
     {
+        /** @var Customer $customer */
         $customer = $this->getUser();
         if (!$customer) {
             return new JsonResponse(['message' => 'Utilisateur non trouvé ou accès interdit'], Response::HTTP_NOT_FOUND);
         }
-        // On récupère tous les utilisateurs
+        // On récupère tous les utilisateurs liés au customer
         $userList = $userRepository->findAllbyCustomer($customer->getId());
         $jsonUserList = $serializer->serialize($userList, 'json', ['groups' => 'getUsers']);
 
@@ -42,13 +43,18 @@ final class UserController extends AbstractController
         SerializerInterface $serializer
         ): JsonResponse
     {
+        /** @var Customer $customer */
         $customer = $this->getUser();
+
         if (!$customer) {
             return new JsonResponse(['message' => 'Utilisateur non trouvé ou accès interdit'], Response::HTTP_NOT_FOUND);
         }
 
         $user = $userRepository->findUserByCustomer($id, $customer->getId());
-
+        if (!$user) {
+            return new JsonResponse(['message' => 'Utilisateur non trouvé ou accès interdit'], Response::HTTP_NOT_FOUND);
+        }
+        
         $jsonUser = $serializer->serialize($user, 'json', ['groups' => 'getUsers']);
         return new JsonResponse($jsonUser, Response::HTTP_OK, [], true);
     }
@@ -59,20 +65,13 @@ final class UserController extends AbstractController
         SerializerInterface $serializer, 
         EntityManagerInterface $em, 
         UrlGeneratorInterface $urlGenerator,
-        CustomerRepository $customerRepository,
         ValidatorInterface $validator
     ): JsonResponse 
     {
         // Désérialisation sans customer
         $user = $serializer->deserialize($request->getContent(), User::class, 'json');
 
-        // Récupérer le customer (l'ID est passé en JSON)
-        $data = json_decode($request->getContent(), true);
-        if (!isset($data['customer']['id'])) {
-            return new JsonResponse(['error' => 'Customer ID is required'], Response::HTTP_BAD_REQUEST);
-        }
-
-        $customer = $customerRepository->find($data['customer']['id']);
+        $customer = $this->getUser();
         if (!$customer) {
             return new JsonResponse(['error' => 'Customer not found'], Response::HTTP_NOT_FOUND);
         }
@@ -81,10 +80,15 @@ final class UserController extends AbstractController
         $user->setCustomer($customer);
 
         // On vérifie les erreurs
-        $errors = $validator->validate($user);
+        $errors = $validator->validate($user); 
 
         if ($errors->count() > 0) {
-            return new JsonResponse($serializer->serialize($errors, 'json'), JsonResponse::HTTP_BAD_REQUEST, [], true);
+            $errorMessages = [];
+            foreach ($errors as $error) {
+                $errorMessages[$error->getPropertyPath()] = $error->getMessage();
+            }
+        
+            return new JsonResponse(['errors' => $errorMessages], JsonResponse::HTTP_BAD_REQUEST);
         }
 
         // Enregistrer en base
@@ -98,43 +102,24 @@ final class UserController extends AbstractController
         return new JsonResponse($jsonUser, Response::HTTP_CREATED, ["Location" => $location], true);
     }
 
-    #[Route('/api/users/{id}', name:"updateUser", methods:['PUT'])]
-
-    public function updateBook(
-        Request $request, 
-        SerializerInterface $serializer, 
-        User $currentUser, 
-        EntityManagerInterface $em, 
-        CustomerRepository $customerRepository
-        ): JsonResponse 
-    {
-        $updatedUser = $serializer->deserialize($request->getContent(), 
-                User::class, 
-                'json', 
-                [AbstractNormalizer::OBJECT_TO_POPULATE => $currentUser]);
-        // Récupérer le customer (l'ID est passé en JSON)
-        $data = json_decode($request->getContent(), true);
-        if (!isset($data['customer']['id'])) {
-            return new JsonResponse(['error' => 'Customer ID is required'], Response::HTTP_BAD_REQUEST);
-        }
-
-        $customer = $customerRepository->find($data['customer']['id']);
-        if (!$customer) {
-            return new JsonResponse(['error' => 'Customer not found'], Response::HTTP_NOT_FOUND);
-        }
-        $updatedUser->setCustomer($customerRepository->find($customer));
-        $jsonUpdateUser = $serializer->serialize($updatedUser, 'json', ['groups' => 'getUsers']);
-        
-        $em->persist($updatedUser);
-        $em->flush();
-        return new JsonResponse($jsonUpdateUser, JsonResponse::HTTP_OK, [], true);
-   }
-
    #[Route('/api/users/{id}', name: 'deleteUser', methods: ['DELETE'])]
-    public function deleteBook(User $user, EntityManagerInterface $em): JsonResponse 
+    public function deleteBook($id, User $user, EntityManagerInterface $em, UserRepository $userRepository, ): JsonResponse 
     {
-        $em->remove($user);
-        $em->flush();
+        /** @var Customer $customer */
+        $customer = $this->getUser();
+
+        if (!$customer) {
+            return new JsonResponse(['message' => 'Utilisateur non trouvé ou accès interdit'], Response::HTTP_NOT_FOUND);
+        }
+
+        $user = $userRepository->findUserByCustomer($id, $customer->getId());
+        if (!$user) {
+            return new JsonResponse(['message' => 'Utilisateur non trouvé ou accès interdit'], Response::HTTP_NOT_FOUND);
+        } else {
+            $em->remove($user);
+            $em->flush();
+        }
+        
 
         return new JsonResponse(null, Response::HTTP_NO_CONTENT);
     }
